Generating symbolic link for the theme itself
Now, you can clone the project straight into your `/wp-content/themes`.
Once you build it, it will create a _symbolic link_ to the theme directory in your project in the parent's directory. With this, users will be able to see the _PWAWP-theme_ as one of the themes in their WP admin.

For the deploy, we should copy the dist diretory and the symbolic link. I believe we can make it better, though.